Clean up, Fix errors and add editor configs

+++ QRGameEngine/Scripting/CSMonoCore.cpp
@@ -332,7 +332,7 @@ bool CSMonoCore::CheckIfMonoMethodExists(const MonoMethodHandle& method_class)
 
 MonoMethodHandle CSMonoCore::TryRegisterMonoMethod(const MonoClassHandle& class_handle, const std::string& method_name)
 {
-	MonoMethodHandle mono_method_handle = { -1 };
+	MonoMethodHandle mono_method_handle = { (uint64_t)(-1) };
 	if (CheckIfMonoMethodExists(class_handle, method_name))
 	{
 		return RegisterMonoMethod(class_handle, method_name);
